user-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
PyJWT==2.10.1
bcrypt==4.2.1
email-validator==2.2.0
redis>=5.0.0,<6
httpx>=0.27.0,<1
google-api-python-client>=2.140.0,<3
google-auth>=2.30.0,<3
google-auth-httplib2>=0.2.0,<1
